renamed components according to conventions
